trace-wallet/vault_analytics/requirements.txt
pywebview>=5.0
pythonnet>=3.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
urllib3>=2.0.0
charset-normalizer>=3.0.0
idna>=3.0
certifi>=2023.0
packaging>=23.0
